requirements/pip.txt
boto3==1.13.5
botocore==1.16.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2.12
django-cors-headers==3.2.1
django-crontab==0.7.1
django-decouple==2.1
django-filter==2.2.0
Django-Integrator==1.1.1.1
django-mathfilters==1.0.0
django-query-signals==0.0.2
django-rest-passwordreset==1.1.0
django-storages==1.9.1
django-tinymce==3.0.2
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
docopt==0.6.2
docutils==0.15.2
html5lib==1.0.1
idna==2.9
jmespath==0.9.5
jwt==1.0.0
Markdown==3.2.1
num2words==0.5.10
numpy==1.18.4
pandas==1.0.3
pdfkit==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.42
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.9
webencodings==0.5.1
xhtml2pdf==0.2.4
xmltodict==0.12.0